moved mc_vector tests back to the src file
